refactor(projectiles): split pools runtime into focused modules
