+++ src/crimson/projectiles/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from .pools import (
+from .runtime import (
     PROJECTILE_BEHAVIOR_BY_TYPE_ID,
     ProjectileBehavior,
     ProjectilePool,
